requirements.txt
pandas
requests
streamlit
snowflake-connector-python
python-dotenv
pycryptodome
snowflake-snowpark-python
streamlit-option-menu